refactor(channelBridges): share filesystem-safe accountId validator
